+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/elements/buttons/MenuSelectShotlockButton.java
@@ -21,6 +21,7 @@ import online.kingdomkeys.kingdomkeys.client.gui.menu.items.equipment.MenuShotlo
 import online.kingdomkeys.kingdomkeys.client.sound.ModSounds;
 import online.kingdomkeys.kingdomkeys.data.PlayerData;
 import online.kingdomkeys.kingdomkeys.item.ShotlockItem;
+import online.kingdomkeys.kingdomkeys.lib.Constants;
 import online.kingdomkeys.kingdomkeys.network.PacketHandler;
 import online.kingdomkeys.kingdomkeys.network.cts.CSEquipShotlock;
 import online.kingdomkeys.kingdomkeys.shotlock.ModShotlocks;
@@ -30,7 +31,6 @@ import online.kingdomkeys.kingdomkeys.util.Utils;
 import java.awt.*;
 
 public class MenuSelectShotlockButton extends MenuButtonBase {
-	final ResourceLocation texture = KingdomKeys.rl("textures/gui/menu/menu_button.png");
 	public int slot;
 	public ItemStack stack;
 	boolean selected;
@@ -84,11 +84,11 @@ public class MenuSelectShotlockButton extends MenuButtonBase {
 			RenderSystem.setShaderColor(col.getRed() / 255F, col.getGreen() / 255F, col.getBlue() / 255F, 1);
 			matrixStack.translate(getX() + 0.6F, getY(), 0);
 			matrixStack.scale(0.5F, 0.5F, 1);
-			gui.blit(texture, 0, 0, 166, 34, 18, 28);
-			gui.blit(texture, 16, 0, ((width * 2) - (17 + 17)) + 2, 28, 186, 34, 2, 28, 256, 256);
-			gui.blit(texture, ((width * 2) - 17), 0, 186, 34, 17, 28);
+			gui.blit(Constants.MENU_TEXTURE, 0, 0, 166, 34, 18, 28);
+			gui.blit(Constants.MENU_TEXTURE, 16, 0, ((width * 2) - (17 + 17)) + 2, 28, 186, 34, 2, 28, 256, 256);
+			gui.blit(Constants.MENU_TEXTURE, ((width * 2) - 17), 0, 186, 34, 17, 28);
 			RenderSystem.setShaderColor(1, 1, 1, 1);
-			gui.blit(texture, 6, 4, category.getU(), category.getV(), 20, 20);
+			gui.blit(Constants.MENU_TEXTURE, 6, 4, category.getU(), category.getV(), 20, 20);
 			matrixStack.popPose();
 
 			String shName = shotlock == null ? "---" : stack.getHoverName().getString();
@@ -104,8 +104,8 @@ public class MenuSelectShotlockButton extends MenuButtonBase {
 				float percent = shotlockItem.getLocalPercent(stack);
 				int barWidth = minecraft.font.width(text);
 				int percentWidth = (int) (barWidth * percent);
-				gui.blit(texture, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, barWidth, 2, 161, 67, 1, 5, 256, 256);
-				gui.blit(texture, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, percentWidth, 2, 163, 67, 1, 5, 256, 256);
+				gui.blit(Constants.MENU_TEXTURE, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, barWidth, 2, 161, 67, 1, 5, 256, 256);
+				gui.blit(Constants.MENU_TEXTURE, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, percentWidth, 2, 163, 67, 1, 5, 256, 256);
 			}
 
 			if (isHovered || selected) {
@@ -114,9 +114,9 @@ public class MenuSelectShotlockButton extends MenuButtonBase {
 					RenderSystem.enableBlend();
 					matrixStack.translate(getX() + 0.6F, getY(), 0);
 					matrixStack.scale(0.5F, 0.5F, 1);
-					gui.blit(texture, 0, 0, 128, 34, 18, 28);
-					gui.blit(texture, 16, 0, ((width * 2) - (17 * 2)) + 2, 28, 148, 34, 2, 28, 256, 256);
-					gui.blit(texture, ((width * 2) - 17), 0, 148, 34, 17, 28);
+					gui.blit(Constants.MENU_TEXTURE, 0, 0, 128, 34, 18, 28);
+					gui.blit(Constants.MENU_TEXTURE, 16, 0, ((width * 2) - (17 * 2)) + 2, 28, 148, 34, 2, 28, 256, 256);
+					gui.blit(Constants.MENU_TEXTURE, ((width * 2) - 17), 0, 148, 34, 17, 28);
 					RenderSystem.disableBlend();
 				}
 				matrixStack.popPose();
@@ -166,8 +166,8 @@ public class MenuSelectShotlockButton extends MenuButtonBase {
 				int barWidth = (int) (parent.boxR.getWidth() * 0.8F);
 				int percentWidth = (int) (barWidth * percent);
 
-				gui.blit(texture, (int) strPosX, (int) posY + 10, barWidth, 5, 161, 67, 1, 5, 256, 256);
-				gui.blit(texture, (int) strPosX, (int) posY + 10, percentWidth, 5, 163, 67, 1, 5, 256, 256);
+				gui.blit(Constants.MENU_TEXTURE, (int) strPosX, (int) posY + 10, barWidth, 5, 161, 67, 1, 5, 256, 256);
+				gui.blit(Constants.MENU_TEXTURE, (int) strPosX, (int) posY + 10, percentWidth, 5, 163, 67, 1, 5, 256, 256);
 			}
 		}
 	}