+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/elements/buttons/MenuStockItem.java
@@ -27,6 +27,7 @@ import online.kingdomkeys.kingdomkeys.data.PlayerData;
 import online.kingdomkeys.kingdomkeys.item.KeybladeItem;
 import online.kingdomkeys.kingdomkeys.item.KeychainItem;
 import online.kingdomkeys.kingdomkeys.item.MagicSpellItem;
+import online.kingdomkeys.kingdomkeys.lib.Constants;
 import online.kingdomkeys.kingdomkeys.synthesis.recipe.Recipe;
 import online.kingdomkeys.kingdomkeys.synthesis.recipe.RecipeRegistry;
 import online.kingdomkeys.kingdomkeys.synthesis.shop.ShopItem;
@@ -48,10 +49,6 @@ public class MenuStockItem extends Button {
     public int offsetY;
     public Color backgroundColor;
     public ChatFormatting textColor = ChatFormatting.WHITE;
-
-    final ResourceLocation texture = KingdomKeys.rl("textures/gui/menu/menu_button.png");
-	final ResourceLocation barTexture = KingdomKeys.rl("textures/gui/menu/menu_button.png");
-
 
 	public MenuStockItem(MenuFilterable parent, ResourceLocation rl, ItemStack displayStack, int x, int y, int width, boolean showAmount, OnPress onPress) {
         super(new Builder(Component.literal(""), onPress).bounds(x, y, width, 14));
@@ -119,9 +116,9 @@ public class MenuStockItem extends Button {
                     matrixStack.translate(getX() + 0.6F, getY(), 0);
                     float scale = 0.5F;
                     matrixStack.scale(scale, scale, 1);
-                    gui.blit(texture, 0, 0, 27, 0, 18, 28);
-                    gui.blit(texture, 16, 0, (int) ((width * (1 / scale)) - (17 * (1 / scale)))+1, 28, 47, 0, 2, 28, 256, 256);
-                    gui.blit(texture, (int)(width * (1 / scale)) - 18, 0, 49, 0, 17, 28);
+                    gui.blit(Constants.MENU_TEXTURE, 0, 0, 27, 0, 18, 28);
+                    gui.blit(Constants.MENU_TEXTURE, 16, 0, (int) ((width * (1 / scale)) - (17 * (1 / scale)))+1, 28, 47, 0, 2, 28, 256, 256);
+                    gui.blit(Constants.MENU_TEXTURE, (int)(width * (1 / scale)) - 18, 0, 49, 0, 17, 28);
                 }
                 matrixStack.popPose();
             } else if(getBackgroundColor() != null) {
@@ -133,9 +130,9 @@ public class MenuStockItem extends Button {
                     float scale = 0.5F;
                     matrixStack.scale(scale, scale, 1);
                     RenderSystem.setShaderColor(getBackgroundColor().getRed()/255F,getBackgroundColor().getGreen()/255F,getBackgroundColor().getBlue()/255F,1);
-                    gui.blit(texture, 0, 0, 219, 0, 18, 28);
-                    gui.blit(texture, 16, 0, (int) ((width * (1 / scale)) - (17 * (1 / scale)))+1, 28, 239, 0, 2, 28, 256, 256);
-                    gui.blit(texture, (int)(width * (1 / scale)) - 18, 0, 239, 0, 17, 28);
+                    gui.blit(Constants.MENU_TEXTURE, 0, 0, 219, 0, 18, 28);
+                    gui.blit(Constants.MENU_TEXTURE, 16, 0, (int) ((width * (1 / scale)) - (17 * (1 / scale)))+1, 28, 239, 0, 2, 28, 256, 256);
+                    gui.blit(Constants.MENU_TEXTURE, (int)(width * (1 / scale)) - 18, 0, 239, 0, 17, 28);
                     RenderSystem.setShaderColor(1,1,1,1);
                 }
                 matrixStack.popPose();
@@ -147,7 +144,7 @@ public class MenuStockItem extends Button {
                 float scale = 0.5F;
                 int categorySize = 20;
                 matrixStack.scale(scale, scale, 1);
-                gui.blit(texture, 0, 0, category.getU(), category.getV(), categorySize, categorySize);
+                gui.blit(Constants.MENU_TEXTURE, 0, 0, category.getU(), category.getV(), categorySize, categorySize);
             }
             matrixStack.popPose();
 
@@ -199,8 +196,8 @@ public class MenuStockItem extends Button {
                     float percent = spell.getLocalPercent(stack);
                     int barWidth = 24;
                     int percentWidth = (int)(barWidth * percent);
-                    gui.blit(barTexture, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, barWidth, 2, 161, 67, 1, 5, 256, 256);
-                    gui.blit(barTexture, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, percentWidth, 2, 163, 67, 1, 5, 256, 256);
+                    gui.blit(Constants.MENU_TEXTURE, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, barWidth, 2, 161, 67, 1, 5, 256, 256);
+                    gui.blit(Constants.MENU_TEXTURE, getX() + getWidth() - barWidth - 5, getY() + getHeight() - 4, percentWidth, 2, 163, 67, 1, 5, 256, 256);
                 }
             }